refactor: Lloyd's re-engineering of API (#7)
